+++ src/service/transaction/refund.ts
@@ -75,6 +75,9 @@ export function start(
             transaction = await repos.transaction.start<factory.transactionType.Refund>(startParams);
 
             switch (paymentServiceType) {
+                case factory.service.paymentService.PaymentServiceType.Account:
+                    break;
+
                 case factory.service.paymentService.PaymentServiceType.CreditCard:
                     break;